Prevent invalid reactions (#20)
I could not fully reproduce this on the irslackd.slack.com
workspace, so I'm not sure this will completely fix our
issue. However, I did see that certain reaction events do
not contain a target user or channel. The docs confirm
this behavior. For now, when we encounter this case, log
the event info and bail.