Implement sending and receiving user_typing events (#47)
* Implement sending and receiving user_typing events

To bridge between IRC and Slack, we rely on the CTCP TYPING command to
send and receive typing notifications.  As with all CTCP commands, these
messages are bundled in a PRIVMSG in the form:

    PRIVMSG \001TYPING LEVEL\001

where LEVEL is an integer (zero means stop typing and non-zero means
typing).

To handle sending user_typing events from the IRC client to Slack, the
server has a check in the onIrcPrivmsg handler for CTCP TYPING commands.
If the PRIVMSG message contains a CTCP TYPING command, the server calls
the sendTyping method of the SlackRtm object to notify Slack that the
specified user in the PRIVMSG is typing.

To handle receiving user_typing events from Slack, the server has a new
onSlackUserTyping handler which sends a CTCP TYPING command with the
specified user to the corresponding IRC channel.

To test the user_typing events, I used Weechat and the following script:

    https://gist.github.com/pbui/92b5a1b3bee87a95331d1f2d714f8f9c

With this new server extension and Weechat can both send typing
notifications and receive them.

Note: Slack does not have a mechanism for clearing typing state, so each
client must keep track of typing state and use some sort of timeout to
clear it as shown in the script above.

Finally, this feature is only enabled if "typing-notifications" is sent
to the server as an argument after the Slack token (similar to
no-threads).

* Add CTCP unit tests

- Add tests for sending and receiving CTCP ACTION commands (ie. /me
messages)

- Add tests for receiving user_typing Slack events and translating those
into CTCP TYPING commands to the IRC client

- We do not have a test for receiving CTCP TYPING commands from the IRC
client and sending a Slack Rtm command because the current
MockSlackRtmClient is not instrumented for testing Rtm commands.