Implement presence monitoring
This feature implements a crude form of user presence by using the CHAN
MODE IRC command to denote if a user is active (+v / voiced) or away (-v
/ devoiced).  The use of voice and devoice to denote presence is based
on how bitlbee communicates status in its IRC bridge.

- To receive presence updates, we must subscribe to each user we are
interested in; unfortunately there is no way to know which channel the
IRC user is viewing, so we simply subscribe to all users.  This
exhaustive approach is expensive, but it is infrequent.

- Likewise, to communicate a change in user presence, we must send a
CHAN MODE command to each channel the user is in.  Again this is
expensive, but it is infrequent and the only way since IRC does not
directly support the concept of user presence.

- No additional bookkeeping is required since the ircUser class already
keeps track of all Slack users in its `slackToIrc` member.  We added a
`IrcUser::getAllSlackUserIds()` method that returns all such Slack user
ids.

- Currently, although batch_presence_aware is set, it appears to do
nothing (presence updates still occur one user at a time).

- On start up, there is a huge flood of traffic since the Slack server
will send updates for many (if not all) of the users (and the bridge
will then send CHAN MODE commands).

- To prevent a race condition where user presence updates come before
the IRC client has joined a client, we add a `ircNickActive` set that
keeps track if an IRC nick is active or not.  The
`onSlackPresenceChange` method either adds the corresponding Slack user
to this set or removes them based on their presence.  The
`ircNickActive` set is also used in the `sendIrcChannelJoin` method to
prepend a `+` to any IRC nicks that are active.

- We add a Slack `team_join` event handler that updates the slackToIrc
map and then updates the presence subscriptions.

- This feature is disabled by default and is only enabled by the
"presence" password argument (sent by the IRC client after the Slack
token).