Add support for "@slack chat" to initiate MPDMs (#94).
This commit implements the suggestion in issue #94 by adding a new
@slack command for initiating new multi-party messaging sessions. With
this change a user can enter:

    @slack chat nick1 nick2 nick3

And a group chat will be initiated and the client automatically force
joined into a #mpdm irc channel.

Naturally '@slack chat nick' and '/query nick' are basically equivalent
in this case.

If the user fails to specify at least one nick, or one of the nicks is
unknown, an appropriate error is returned.