Remove numeric suffix from short group chat names.
I noticed this little glitch in the last push but didn't think it was a
wart worth fixing, but then I noticed something amusing:  the numeric
suffixed (i.e. "-1") becomes part of the last nick name.  Well, if your
own nickname is the last nick in the channel name, then it doesn't get
pulled off, which means the abbreviation isn't complete.  Oops.

This changeset tweaks this behaviour so the numeric suffix is also
pulled off, which then corrects this knock-on effect.

TBH, I'm not sure why Slack puts that "-1" at the end of mpdm's...
unless you can have more than one with the same group of people
simultaneously?  Seems unlikely, though...